move kustoOptions to common package
